refactor(generation): split GenerationView into focused extension files
Carve the 1,900-line view into GenerationView+ModelState/References/
Mentions/Settings/Submit extensions, and extract RefCard, RefDropZone,
and FrameSlot as standalone views. Verbatim moves; no behavior change.
